Habilidades Especiales

+++ src/Videojuego/Personaje.java
@@ -5,7 +5,7 @@ public abstract class Personaje {
     public double Vida;
     public double Ataque;
     public double Defensa;
-    public int Velocidad;
+    public double Velocidad;
     public double danoIngfligido = 0;
     public String nombre;
     public boolean Nerffeado=false;
